requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
groq==0.9.0
aiofiles==23.2.1
aiohttp==3.9.1
pydub==0.25.1
requests==2.31.0
uuid==1.30
soundfile==0.12.1
python-magic-bin==0.4.14